requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiogram-dialog==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
dynaconf==3.2.5
eventlet==0.36.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
logging==0.4.9.6
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
ormsgpack==1.5.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
SQLAlchemy==2.0.20
structlog==24.2.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
